delete unused components
