--- app/(components)/Column/CardColumn.css
@@ -1,11 +1,0 @@
-.column {
-  background-color: #c9c9f0;
-  border-radius: 5px;
-  border-width: 0px;
-  padding: 15px;
-  width: 80%;
-  max-width: 300px;
-  display: flex;
-  flex-direction: column;
-  gap: 15px;
-}
