--- app/(components)/Input/Input.css
@@ -1,10 +1,0 @@
-.container {
-  display: flex;
-  gap: 10px;
-}
-
-.input {
-  border: 2px solid #ddd;
-  border-radius: 10px;
-  padding: 10px;
-}
