+++ app/(components)/TaskCard/TaskCard.jsx
@@ -1,70 +1,119 @@
-"use client";
-
-import React from "react";
-import "./TaskCard.css";
+import { useState } from "react";
 import { useSortable } from "@dnd-kit/sortable";
 import { CSS } from "@dnd-kit/utilities";
+import TrashIcon from "@/app/Icons/TrashIcon";
+import PropTypes from "prop-types";
 
-import UserIcon from "../../SVG/UserIcon";
-import CalendarIcon from "../../SVG/CalendarIcon";
-import CheckIcon from "../../SVG/CheckIcon";
+function TaskCard({ task, deleteTask, updateTask }) {
+  const [mouseIsOver, setMouseIsOver] = useState(false);
+  const [editMode, setEditMode] = useState(false);
 
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
+  // Make tasks sortable
+  const {
+    setNodeRef,
+    attributes,
+    listeners,
+    transform,
+    transition,
+    isDragging,
+  } = useSortable({
+    id: task.id,
+    data: {
+      type: "Task",
+      task,
+    },
+    disabled: editMode, // Disable Drag+Drop when editing Card Title
+  });
 
-export const TaskCard = ({ id, title, description, content, footer }) => {
-  const { attributes, listeners, setNodeRef, transform, transition } =
-    useSortable({ id });
-
-  // * Look at DND-Kit Tips: I believe this is where -> Converting the `transform` object to a string can feel tedious. Fear not, you can avoid having to do this by hand by importing the `CSS` utility from the `@dnd-kit/utilities` package:
-  // This should be within your component that receives `transform` from `useDraggable`:
-  // 'style' Controls what the element looks like while dragging it
   const style = {
-    transition, // <- from youtube tutorial
-    // transform: CSS.Transform.toString(transform), // <- from youtube tutorial
-    transform: CSS.Translate.toString(transform), // <- from notes/official tips
+    transition,
+    transform: CSS.Translate.toString(transform),
   };
 
+  const toggleEditMode = () => {
+    setEditMode((prev) => !prev);
+    setMouseIsOver(false);
+  };
+
+  // Returns TaskCard slot where being dragged from, then dragged to
+  if (isDragging) {
+    // bg-mainBackgroundColor
+    return (
+      <div
+        className="task relative flex h-[100px] min-h-[100px] cursor-grab items-center rounded-xl border-2 border-rose-500 bg-green-600 p-2.5 text-left opacity-50"
+        ref={setNodeRef}
+        style={style}
+      />
+    );
+  }
+
+  // Returned when IN EDIT MODE
+  if (editMode) {
+    return (
+      <div
+        className="relative flex h-[100px] min-h-[100px] cursor-grab items-center rounded-xl bg-blue-900 p-2.5 text-left hover:ring-2 hover:ring-inset hover:ring-rose-500"
+        ref={setNodeRef}
+        style={style}
+        {...attributes}
+        {...listeners}
+      >
+        <textarea
+          className="h-[90%] w-full resize-none rounded border-none bg-transparent text-white focus:outline-none"
+          value={task.content}
+          autoFocus
+          placeholder="Task content here. Press Shift+Enter to Submit"
+          onBlur={toggleEditMode}
+          onKeyDown={(e) => {
+            if (e.key === "Enter" && e.shiftKey) {
+              toggleEditMode();
+            }
+          }}
+          onChange={(e) => updateTask(task.id, e.target.value)}
+        />
+      </div>
+    );
+  }
+
+  // This is returned when NOT IN EDIT MODE (Standard)
   return (
     <div
+      className="task relative flex h-[100px] min-h-[100px] cursor-grab items-center rounded-xl bg-red-800 p-2.5 text-left hover:ring-2 hover:ring-inset hover:ring-rose-500"
       ref={setNodeRef}
+      style={style}
       {...attributes}
       {...listeners}
-      style={style}
-      className="task-card"
+      onClick={toggleEditMode}
+      onMouseEnter={() => {
+        setMouseIsOver(true);
+      }}
+      onMouseLeave={() => {
+        setMouseIsOver(false);
+      }}
     >
-      <Card>
-        <CardHeader>
-          {/* Checkbox proves that check state is remembered after card shuffle */}
-          {/* <input type="checkbox" className="checkbox" /> */}
-          <CardTitle>{title}</CardTitle>
-          <CardDescription>{description}</CardDescription>
-          <CardDescription>ID: {id}</CardDescription>
-        </CardHeader>
-        {/* !!! DISABLED TEMPORARILY WHILE I TEST CARD COLUMN SORT */}
-        {/* <CardContent>
-          <p>{content}</p>
-          <div className="flex items-center space-x-2">
-            <UserIcon className="h-5 w-5 text-gray-500 dark:text-gray-400" />
-            <span className="text-sm">Assignee: John Doe</span>
-          </div>
-          <div className="flex items-center space-x-2">
-            <CalendarIcon className="h-5 w-5 text-gray-500 dark:text-gray-400" />
-            <span className="text-sm">Due Date: February 20, 2024</span>
-          </div>
-          <div className="flex items-center space-x-2">
-            <CheckIcon className="h-5 w-5 text-gray-500 dark:text-gray-400" />
-            <span className="text-sm">Status: In Progress</span>
-          </div>
-        </CardContent>
-        <CardFooter>{footer}</CardFooter> */}
-      </Card>
+      <p className="my-auto h-[90%] w-full overflow-y-auto overflow-x-hidden whitespace-pre-wrap">
+        {task.content}
+      </p>
+      {mouseIsOver && (
+        <button
+          className="top-1/2-translate-y-1/2 absolute right-4 rounded bg-columnBackgroundColor stroke-white p-2 opacity-60 hover:opacity-100"
+          onClick={() => {
+            deleteTask(task.id);
+          }}
+        >
+          <TrashIcon />
+        </button>
+      )}
     </div>
   );
+}
+
+export default TaskCard;
+
+TaskCard.propTypes = {
+  task: PropTypes.shape({
+    id: PropTypes.string.isRequired,
+    content: PropTypes.string.isRequired,
+  }),
+  deleteTask: PropTypes.func.isRequired,
+  updateTask: PropTypes.func.isRequired,
 };
