--- app/Board/Draggable.jsx
@@ -1,24 +1,0 @@
-// ! Currently Unused
-
-// "use client";
-
-import React from "react";
-import { useDraggable } from "@dnd-kit/core";
-
-// Added 'export'
-export function Draggable(props) {
-  const { attributes, listeners, setNodeRef, transform } = useDraggable({
-    id: "draggable",
-  });
-  const style = transform
-    ? {
-        transform: `translate3d(${transform.x}px, ${transform.y}px, 0)`,
-      }
-    : undefined;
-
-  return (
-    <button ref={setNodeRef} style={style} {...listeners} {...attributes}>
-      {props.children}
-    </button>
-  );
-}
